requirements.txt
absl-py==1.0.0
anyio==3.2.1
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
blis==0.7.5
cachetools==4.2.4`
catalogue==2.0.6
certifi==2021.10.8
cffi==1.14.5
chardet==4.0.0
click==8.0.3
cycler==0.10.0
cymem==2.0.6
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.9
django-cors-headers==3.10.1
editdistance==0.6.0
efficientnet==1.0.0
entrypoints==0.3
essential-generators==1.0
fastai==2.5.3
fastcore==1.3.27
fastdownload==0.0.5
fastprogress==1.0.0
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
gTTS==2.2.3
gunicorn==20.1.0
h5py==3.6.0
idna==2.10
image==1.5.33
imageio==2.13.3
imgaug==0.4.0
importlib-metadata==4.8.2
imutils==0.5.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Applications==1.0.8
keras-models==0.0.7
keras-ocr==0.8.9
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.1
langcodes==3.3.0
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
MouseInfo==0.1.3
murmurhash==1.0.6
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.0
npm==0.1.1
npx==0.0.24
numpy==1.21.1
oauthlib==3.1.1
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opt-einsum==3.3.0
optional-django==0.1.0
packaging==20.9
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pathlib==1.0.1
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
playsound==1.3.0
preshed==3.0.6
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.8.2
pydotplus==2.0.2
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.9.0
PyMsgBox==1.0.9
pyobjc==7.3
pyobjc-core==7.3
pyobjc-framework-Accessibility==7.3
pyobjc-framework-Accounts==7.3
pyobjc-framework-AddressBook==7.3
pyobjc-framework-AdServices==7.3
pyobjc-framework-AdSupport==7.3
pyobjc-framework-AppleScriptKit==7.3
pyobjc-framework-AppleScriptObjC==7.3
pyobjc-framework-ApplicationServices==7.3
pyobjc-framework-AppTrackingTransparency==7.3
pyobjc-framework-AuthenticationServices==7.3
pyobjc-framework-AutomaticAssessmentConfiguration==7.3
pyobjc-framework-Automator==7.3
pyobjc-framework-AVFoundation==7.3
pyobjc-framework-AVKit==7.3
pyobjc-framework-BusinessChat==7.3
pyobjc-framework-CalendarStore==7.3
pyobjc-framework-CallKit==7.3
pyobjc-framework-CFNetwork==7.3
pyobjc-framework-ClassKit==7.3
pyobjc-framework-CloudKit==7.3
pyobjc-framework-Cocoa==7.3
pyobjc-framework-Collaboration==7.3
pyobjc-framework-ColorSync==7.3
pyobjc-framework-Contacts==7.3
pyobjc-framework-ContactsUI==7.3
pyobjc-framework-CoreAudio==7.3
pyobjc-framework-CoreAudioKit==7.3
pyobjc-framework-CoreBluetooth==7.3
pyobjc-framework-CoreData==7.3
pyobjc-framework-CoreHaptics==7.3
pyobjc-framework-CoreLocation==7.3
pyobjc-framework-CoreMedia==7.3
pyobjc-framework-CoreMediaIO==7.3
pyobjc-framework-CoreMIDI==7.3
pyobjc-framework-CoreML==7.3
pyobjc-framework-CoreMotion==7.3
pyobjc-framework-CoreServices==7.3
pyobjc-framework-CoreSpotlight==7.3
pyobjc-framework-CoreText==7.3
pyobjc-framework-CoreWLAN==7.3
pyobjc-framework-CryptoTokenKit==7.3
pyobjc-framework-DeviceCheck==7.3
pyobjc-framework-DictionaryServices==7.3
pyobjc-framework-DiscRecording==7.3
pyobjc-framework-DiscRecordingUI==7.3
pyobjc-framework-DiskArbitration==7.3
pyobjc-framework-DVDPlayback==7.3
pyobjc-framework-EventKit==7.3
pyobjc-framework-ExceptionHandling==7.3
pyobjc-framework-ExecutionPolicy==7.3
pyobjc-framework-ExternalAccessory==7.3
pyobjc-framework-FileProvider==7.3
pyobjc-framework-FileProviderUI==7.3
pyobjc-framework-FinderSync==7.3
pyobjc-framework-FSEvents==7.3
pyobjc-framework-GameCenter==7.3
pyobjc-framework-GameController==7.3
pyobjc-framework-GameKit==7.3
pyobjc-framework-GameplayKit==7.3
pyobjc-framework-ImageCaptureCore==7.3
pyobjc-framework-IMServicePlugIn==7.3
pyobjc-framework-InputMethodKit==7.3
pyobjc-framework-InstallerPlugins==7.3
pyobjc-framework-InstantMessage==7.3
pyobjc-framework-Intents==7.3
pyobjc-framework-IOSurface==7.3
pyobjc-framework-iTunesLibrary==7.3
pyobjc-framework-KernelManagement==7.3
pyobjc-framework-LatentSemanticMapping==7.3
pyobjc-framework-LaunchServices==7.3
pyobjc-framework-libdispatch==7.3
pyobjc-framework-LinkPresentation==7.3
pyobjc-framework-LocalAuthentication==7.3
pyobjc-framework-MapKit==7.3
pyobjc-framework-MediaAccessibility==7.3
pyobjc-framework-MediaLibrary==7.3
pyobjc-framework-MediaPlayer==7.3
pyobjc-framework-MediaToolbox==7.3
pyobjc-framework-Metal==7.3
pyobjc-framework-MetalKit==7.3
pyobjc-framework-MetalPerformanceShaders==7.3
pyobjc-framework-MetalPerformanceShadersGraph==7.3
pyobjc-framework-MLCompute==7.3
pyobjc-framework-ModelIO==7.3
pyobjc-framework-MultipeerConnectivity==7.3
pyobjc-framework-NaturalLanguage==7.3
pyobjc-framework-NetFS==7.3
pyobjc-framework-Network==7.3
pyobjc-framework-NetworkExtension==7.3
pyobjc-framework-NotificationCenter==7.3
pyobjc-framework-OpenDirectory==7.3
pyobjc-framework-OSAKit==7.3
pyobjc-framework-OSLog==7.3
pyobjc-framework-PassKit==7.3
pyobjc-framework-PencilKit==7.3
pyobjc-framework-Photos==7.3
pyobjc-framework-PhotosUI==7.3
pyobjc-framework-PreferencePanes==7.3
pyobjc-framework-PushKit==7.3
pyobjc-framework-Quartz==7.3
pyobjc-framework-QuickLookThumbnailing==7.3
pyobjc-framework-ReplayKit==7.3
pyobjc-framework-SafariServices==7.3
pyobjc-framework-SceneKit==7.3
pyobjc-framework-ScreenSaver==7.3
pyobjc-framework-ScreenTime==7.3
pyobjc-framework-ScriptingBridge==7.3
pyobjc-framework-SearchKit==7.3
pyobjc-framework-Security==7.3
pyobjc-framework-SecurityFoundation==7.3
pyobjc-framework-SecurityInterface==7.3
pyobjc-framework-ServiceManagement==7.3
pyobjc-framework-Social==7.3
pyobjc-framework-SoundAnalysis==7.3
pyobjc-framework-Speech==7.3
pyobjc-framework-SpriteKit==7.3
pyobjc-framework-StoreKit==7.3
pyobjc-framework-SyncServices==7.3
pyobjc-framework-SystemConfiguration==7.3
pyobjc-framework-SystemExtensions==7.3
pyobjc-framework-UniformTypeIdentifiers==7.3
pyobjc-framework-UserNotifications==7.3
pyobjc-framework-UserNotificationsUI==7.3
pyobjc-framework-VideoSubscriberAccount==7.3
pyobjc-framework-VideoToolbox==7.3
pyobjc-framework-Virtualization==7.3
pyobjc-framework-Vision==7.3
pyobjc-framework-WebKit==7.3
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.17.3
PyScreeze==0.1.27
pytesseract==0.3.8
python-dateutil==2.8.1
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.1.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
rubicon-objc==0.4.0
scikit-image==0.19.0
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.7.1
Shapely==1.8.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
sqlparse==0.4.2
srsly==2.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.10.1
tesseract==0.1.3
testpath==0.5.0
thinc==8.0.13
threadpoolctl==2.2.0
tifffile==2021.11.2
toml==0.10.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.5
validators==0.18.2
voila==0.3.0
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
websockets==10.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.6.0
